+++ code/game/machinery/machinery.dm
@@ -535,7 +535,7 @@ Class Procs:
 		if((. = .(candidate)))
 			return
 
-/obj/machinery/proc/on_user_login(mob/M)
+/obj/proc/on_user_login(mob/M)
 	return
 
 /obj/machinery/proc/set_emergency_state(var/new_security_level)
